Performance test commit
This commit shows the (rough) timing result of three different areas:
1. Time spend on networking
2. Time spend on parsing
3. Time spend on writing to disk

Timings (from 'asynctestoutput.txt'):

Total time spend on network: 134.13417047198163
Total time spend on parsing: 6.338895042979857
Total time spend on fileIO: 0.010557281988440081
Total time spend: 140.48944479599595

As seen here a lot of the runtime is spend on networking. Therefore a
runtime speedup could be archived by making the networking part
asynchronous.
Parsing takes not that much time, but could benefit from being run async
next to networking (eg. parsing while the http-requests wait for
answers)
Writing to disk takes a negligible amount of runtime on a SATA-SSD, and
therefore is not a good target for asynchronocity.

As a result further experimentation with making the codebase
asynchronous regarding processing and networking will take place on this
branch.