Make the timer async-compatible
When the timer is started with the start_multi-method it will generate a
new random id for that timer. The end_multi-method now takes this id as
an argument to measure the time-span associated with that specific id.
The start- and end-methods now use the same system, except that their
timer has the fixed (and reserved) id of 0.