Use reworked dprint-function for messaging
The author discovered that this type functionality is normally taken
care of by a logging framework. It will be replaced by such in one of
the following commits. The work is done however so the code will be
saved into the git-history regardless. Maybe to serve as a testimonial
to the fact that an hour of work can save five minutes of googling.

If a user only wants error messages for example (eg. lower verbosity
levels) and the error message needs context, that would only appear
on higher verbosity levels, there now is the possibility of providing
those error printouts with a context-tuple. If the verbosity level would
have been too low for the error message to appear, dprint can now print
it regardless on a case-by-case basis.
Example: One dprint-call shows the currently processed URL and the other
shows a specific error that occurs while processing the URL. If only
error-messaging was allowed the error message would appear without the
URL that was being processed while this error occured, thus rendering
the error message pointless. Through context the error message has
the option to show the relevant information regardless of the verbosity
levels.