Restructure the program
'rezepte.txt' has now been divided into multiple modules across two
packages. The general functionality stays the same, but there have been
a lot of changes to the outsourced functions in order to get them
working outside of their original module (which also makes this commit
kind of messy).
There now exist a module called 're2txt.py', which serves as an entry
point to the program and hopefully can 'predigest' the command line
parameters for the rest of the program in the future.