Change fetcher into a class
fetcher.py (former rezepte.py) is now a class. This ensures that the
constructor can rule out invalid states of the program. It also moves
more of the preprocessing (e.g. 'process-urls()') into an earlier stage
of the program (into r2txt.py), further carving out the responsibility
of the fetcher-module (and therefore class) as a coordinator of async
requests and processing of the results. Preprocessing raw strings (for
the fetcher module) is now firmly in the territory of the r2txt-module.