Reorder attributes in Recipe, add "scraper_version"
The order of attributes in Recipe now reflects the order of
corresponding strings in recipe_attributes. This allows for
initialization from lists or tuples that have been generated by
iterating over recipe_attributes.
This also strongly links the order of Recipe and the order of
recipe_attributes. Changing the order and adding or removing attributes
in Recipes has to be reflected in recipe_attributes (or vice versa).
The injection between the elements of the two cannot be violated.
Failure to comply will lead to (potentially fatal) bugs!