Add more detailed cache handling
There are three ways to deal with cached recipes now:

default: If a recipe is not or not fully cached it is downloaded from
the web. The freshly aquired version is merged with the old data (if
available) and written into the cache-database.

ignore: The cache is ignored. Everything is downloaded regardless of
local availablility. New information will still be written into the
cache.

only: Nothing is downloaded. Everything is fetched from cache. If a
recipe cannot be found in the cache it will not be written into the
final file.