Rewrite set_default_output
The function now makes use of pathlib and logging. A debug-parameter has
been added for easier debugging. When creating a config-file, the
function now writes the same path with two different endings into the
file. In the future there should be some basic checks on the written
paths. Additionally there should be the possibility of different paths
for different endings.