Add the option to attach exceptions to log-messages
The message will either be just the error itself or a full-blown stack-
trace, if log-level is set to debug. The stack trace itself will be
indented as to make the structure of the log-message visually easier to
digest. To avoid too much line-wrapping with long paths, the stack trace
contains one full path, while all other paths have been shortened.
Their first pathelement is the last pathelement that is shared by all
paths in that trace.