Change file initialization order
Since the database file and the log file
will be written to directories that should
always be accessible, the output file has a
higher probability of landing somewhere it
cannot. So to fail as early as possible the
initialization is done before the other two.

Hopefully this commit can serve as an
excellent example of bike-shedding for all
future generations.