Start splitting file_setup-functionality
...into the sub-functions get_db, get_log. This makes the
file-handling-interface more intuitive (instead of having a
function that returns three parameters) and allows the
different files to be initialized at different stages.
For example: The logger needs the log-file. If this file
gets validated first, logging can start earlier in the
program-environment-validation-process, thus allowing for
better feedback about what is happening.

Additionally, in the future the function file_setup would
have had to handle gathering of jinja2-templates as well,
making it even more awkward as an interface. Now an easy-to-
understand-call to get_templates will be far more intuitive
than get_files[3].