Restructure the program-initialisation-process
The program-initialisation now takes place in main(), which calls
multiple smaller function/constructors to get everything in order. A lot
of this comes from the cannibalized process_params-function.

The added functions are smaller than process_params and more focused in
their scope. This should make the initialisation more transparent and
easier to grasp (and dare I say the evil worlds: "self-documenting"?).